✨ Add PlannerCommand and integrate with root command
This commit introduces a new PlannerCommand class that handles the
planning functionality. It has been integrated with the root command,
replacing the previous implementation. The PlannerCommand class now
generates a CreatePlan subcommand and handles its execution. The
RunCreatePlan method has been moved into the PlannerCommand class,
simplifying the Program.cs file.
